move some non-DB-dependent tests to HttpServiceIntegrationTest
CHANGELOG_BEGIN
CHANGELOG_END
