+++ ledger-service/http-json/src/it/scala/http/HttpServiceIntegrationTest.scala
@@ -5,10 +5,15 @@ package com.daml.http
 
 import java.io.File
 import java.nio.file.Files
+import java.util.UUID
 
 import akka.http.scaladsl.Http
 import akka.http.scaladsl.model.{HttpMethods, HttpRequest, StatusCodes, Uri}
-import com.daml.http.dbbackend.JdbcConfig
+import AbstractHttpServiceIntegrationTestFuns.HttpServiceTestFixtureData
+import dbbackend.JdbcConfig
+import json.JsonError
+import util.Logging.instanceUUIDLogCtx
+import com.daml.ledger.api.refinements.{ApiTypes => lar}
 import com.daml.ledger.api.v1.{value => v}
 import com.daml.lf.data.Ref
 import com.daml.lf.value.test.TypedValueGenerators.{ValueAddend => VA}
@@ -17,16 +22,23 @@ import json.SprayJson.{decode => jdecode}
 import com.daml.http.util.TestUtil.writeToFile
 import org.scalacheck.Gen
 import org.scalatest.{Assertion, BeforeAndAfterAll}
-import scalaz.\/-
+import scalaz.{-\/, \/-, EitherT, \/}
+import scalaz.std.scalaFuture._
+import scalaz.syntax.apply._
+import scalaz.syntax.bifunctor._
+import scalaz.syntax.show._
+import scalaz.syntax.tag._
 import shapeless.record.{Record => ShRecord}
 import spray.json.JsValue
 
 import scala.concurrent.Future
 
+@SuppressWarnings(Array("org.wartremover.warts.NonUnitStatements"))
 abstract class HttpServiceIntegrationTest
     extends AbstractHttpServiceIntegrationTestTokenIndependent
     with BeforeAndAfterAll {
   import HttpServiceIntegrationTest._
+  import json.JsonProtocol._
   import AbstractHttpServiceIntegrationTestFuns.ciouDar
 
   private val staticContent: String = "static"
@@ -58,6 +70,202 @@ abstract class HttpServiceIntegrationTest
     super.afterAll()
   }
 
+  "query with invalid JSON query should return error" in withHttpService { fixture =>
+    fixture
+      .postJsonStringRequest(Uri.Path("/v1/query"), "{NOT A VALID JSON OBJECT")
+      .parseResponse[JsValue]
+      .map(inside(_) { case domain.ErrorResponse(_, _, StatusCodes.BadRequest, _) =>
+        succeed
+      }): Future[Assertion]
+  }
+
+  "create" - {
+    import domain.ContractTypeId.OptionalPkg
+
+    "succeeds with single party, proper argument" in withHttpService { fixture =>
+      import fixture.encoder
+      fixture.getUniquePartyAndAuthHeaders("Alice").flatMap { case (alice, headers) =>
+        val command: domain.CreateCommand[v.Record, OptionalPkg] = iouCreateCommand(alice)
+
+        postCreateCommand(command, fixture, headers)
+          .map(inside(_) { case domain.OkResponse(activeContract, _, StatusCodes.OK) =>
+            assertActiveContract(activeContract)(command, encoder)
+          }): Future[Assertion]
+      }
+    }
+
+    // TEST_EVIDENCE: Authorization: reject requests with missing auth header
+    "fails if authorization header is missing" in withHttpService { fixture =>
+      import fixture.encoder
+      val alice = getUniqueParty("Alice")
+      val command: domain.CreateCommand[v.Record, OptionalPkg] = iouCreateCommand(alice)
+      val input: JsValue = encoder.encodeCreateCommand(command).valueOr(e => fail(e.shows))
+
+      fixture
+        .postJsonRequest(Uri.Path("/v1/create"), input, List())
+        .parseResponse[JsValue]
+        .map(inside(_) { case domain.ErrorResponse(Seq(error), _, StatusCodes.Unauthorized, _) =>
+          error should include(
+            "missing Authorization header with OAuth 2.0 Bearer Token"
+          )
+        }): Future[Assertion]
+    }
+
+    "supports extra readAs parties" in withHttpService { fixture =>
+      import fixture.encoder
+      val alice = getUniqueParty("Alice")
+      val command: domain.CreateCommand[v.Record, OptionalPkg] = iouCreateCommand(alice)
+      val input: JsValue = encoder.encodeCreateCommand(command).valueOr(e => fail(e.shows))
+
+      fixture
+        .headersWithPartyAuth(actAs = List(alice.unwrap), readAs = List("Bob"))
+        .flatMap(
+          fixture
+            .postJsonRequest(
+              Uri.Path("/v1/create"),
+              input,
+              _,
+            )
+            .parseResponse[domain.ActiveContract[JsValue]]
+        )
+        .map(inside(_) { case domain.OkResponse(activeContract, _, StatusCodes.OK) =>
+          assertActiveContract(activeContract)(command, encoder)
+        }): Future[Assertion]
+    }
+
+    "with unsupported templateId should return proper error" in withHttpService { fixture =>
+      import fixture.encoder
+      fixture.getUniquePartyAndAuthHeaders("Alice").flatMap { case (alice, headers) =>
+        val command: domain.CreateCommand[v.Record, OptionalPkg] =
+          iouCreateCommand(alice).copy(templateId = domain.TemplateId(None, "Iou", "Dummy"))
+        val input: JsValue = encoder.encodeCreateCommand(command).valueOr(e => fail(e.shows))
+
+        fixture
+          .postJsonRequest(Uri.Path("/v1/create"), input, headers)
+          .parseResponse[JsValue]
+          .map(inside(_) { case domain.ErrorResponse(Seq(error), _, StatusCodes.BadRequest, _) =>
+            val unknownTemplateId: OptionalPkg =
+              domain
+                .TemplateId(None, command.templateId.moduleName, command.templateId.entityName)
+            error should include(
+              s"Cannot resolve template ID, given: ${unknownTemplateId: OptionalPkg}"
+            )
+          }): Future[Assertion]
+      }
+    }
+
+    "supports command deduplication" in withHttpService { fixture =>
+      import fixture.encoder
+      def genSubmissionId() = domain.SubmissionId(UUID.randomUUID().toString)
+      fixture.getUniquePartyAndAuthHeaders("Alice").flatMap { case (alice, headers) =>
+        val cmdId = domain.CommandId apply UUID.randomUUID().toString
+        def cmd(
+            submissionId: domain.SubmissionId
+        ): domain.CreateCommand[v.Record, OptionalPkg] =
+          iouCreateCommand(
+            alice,
+            amount = "19002.0",
+            meta = Some(
+              domain.CommandMeta(
+                commandId = Some(cmdId),
+                actAs = None,
+                readAs = None,
+                submissionId = Some(submissionId),
+                deduplicationPeriod = Some(domain.DeduplicationPeriod.Duration(10000L)),
+              )
+            ),
+          )
+
+        val firstCreate: JsValue =
+          encoder.encodeCreateCommand(cmd(genSubmissionId())).valueOr(e => fail(e.shows))
+
+        fixture
+          .postJsonRequest(Uri.Path("/v1/create"), firstCreate, headers)
+          .parseResponse[domain.CreateCommandResponse[JsValue]]
+          .map(inside(_) { case domain.OkResponse(result, _, _) =>
+            result.completionOffset.unwrap should not be empty
+          })
+          .flatMap { _ =>
+            val secondCreate: JsValue =
+              encoder.encodeCreateCommand(cmd(genSubmissionId())).valueOr(e => fail(e.shows))
+            fixture
+              .postJsonRequest(Uri.Path("/v1/create"), secondCreate, headers)
+              .map(inside(_) { case (StatusCodes.Conflict, _) =>
+                succeed
+              }): Future[Assertion]
+          }
+      }
+    }
+  }
+
+  "create-and-exercise IOU_Transfer" in withHttpService { fixture =>
+    import fixture.encoder
+    fixture.getUniquePartyAndAuthHeaders("Alice").flatMap { case (alice, headers) =>
+      val cmd
+          : domain.CreateAndExerciseCommand[v.Record, v.Value, domain.ContractTypeId.OptionalPkg] =
+        iouCreateAndExerciseTransferCommand(alice)
+
+      val json: JsValue = encoder.encodeCreateAndExerciseCommand(cmd).valueOr(e => fail(e.shows))
+
+      fixture
+        .postJsonRequest(Uri.Path("/v1/create-and-exercise"), json, headers)
+        .parseResponse[domain.ExerciseResponse[JsValue]]
+        .flatMap(inside(_) { case domain.OkResponse(result, None, StatusCodes.OK) =>
+          result.completionOffset.unwrap should not be empty
+          inside(result.events) {
+            case List(
+                  domain.Contract(\/-(created0)),
+                  domain.Contract(-\/(archived0)),
+                  domain.Contract(\/-(created1)),
+                ) =>
+              assertTemplateId(created0.templateId, cmd.templateId)
+              assertTemplateId(archived0.templateId, cmd.templateId)
+              archived0.contractId shouldBe created0.contractId
+              assertTemplateId(created1.templateId, TpId.Iou.IouTransfer)
+              asContractId(result.exerciseResult) shouldBe created1.contractId
+          }
+        })
+    }: Future[Assertion]
+  }
+
+  "should be able to serialize and deserialize domain commands" in withHttpService { fixture =>
+    (testCreateCommandEncodingDecoding(fixture) *>
+      testExerciseCommandEncodingDecoding(fixture)): Future[Assertion]
+  }
+
+  private def testCreateCommandEncodingDecoding(
+      fixture: HttpServiceTestFixtureData
+  ): Future[Assertion] = instanceUUIDLogCtx { implicit lc =>
+    import fixture.{uri, encoder, decoder}
+    import util.ErrorOps._
+    import com.daml.jwt.domain.Jwt
+
+    val command0: domain.CreateCommand[v.Record, domain.ContractTypeId.OptionalPkg] =
+      iouCreateCommand(domain.Party("Alice"))
+
+    type F[A] = EitherT[Future, JsonError, A]
+    val x: F[Assertion] = for {
+      jsVal <- EitherT.either(
+        encoder.encodeCreateCommand(command0).liftErr(JsonError)
+      ): F[JsValue]
+      command1 <- (EitherT.rightT(jwt(uri)): F[Jwt])
+        .flatMap(decoder.decodeCreateCommand(jsVal, _, fixture.ledgerId))
+    } yield command1.bimap(removeRecordId, removePackageId) should ===(command0)
+
+    (x.run: Future[JsonError \/ Assertion]).map(_.fold(e => fail(e.shows), identity))
+  }
+
+  private def testExerciseCommandEncodingDecoding(
+      fixture: HttpServiceTestFixtureData
+  ): Future[Assertion] = {
+    import fixture.{uri, encoder, decoder}
+    val command0 = iouExerciseTransferCommand(lar.ContractId("#a-contract-ID"))
+    val jsVal: JsValue = encodeExercise(encoder)(command0)
+    val command1 =
+      jwt(uri).flatMap(decodeExercise(decoder, _, fixture.ledgerId)(jsVal))
+    command1.map(_.bimap(removeRecordId, identity) should ===(command0))
+  }
+
   "should serve static content from configured directory" in withHttpService {
     (uri: Uri, _, _, _) =>
       Http()
@@ -77,7 +285,6 @@ abstract class HttpServiceIntegrationTest
   }
 
   "exercise interface choices" - {
-    import json.JsonProtocol._
     import AbstractHttpServiceIntegrationTestFuns.{UriFixture, EncoderFixture}
 
     def createIouAndExerciseTransfer(
@@ -224,7 +431,6 @@ abstract class HttpServiceIntegrationTest
 
   "fail to exercise by key with interface ID" in withHttpService { fixture =>
     import fixture.encoder
-    import json.JsonProtocol._
     for {
       _ <- uploadPackage(fixture)(ciouDar)
       aliceH <- fixture.getUniquePartyAndAuthHeaders("Alice")
